Create unzip-merge-output.sh
A script that unzips and grabs the timezone from device status and adds that, plus BG and timestamp data, into a new csv file. It also adds headers for other columns of data that will be added after this process. If needed, I have formulas created in excel to calculate if data is from a control or intervention period or neither; the hour of day the data is from to calculate if it is day or nighttime; and also (once looping start date manually added to file) can calculate number of days looping and number of days of data in the upload to calculate the control/intervention time frames based on the project protocol.