Create devicestatus-pull-isf-timestamp.sh
My first attempt at writing a bash script, so it's not the most efficient. This script assumes that the unzipped and csv-fied files exist. We could add that script above this one if desired. I used csvkit to pull the ISF and timestamp variables. I'd like to avoid creating two separate csv files and then deleting them, but I was struggling with calling only ISF and timestamp with just one line of code, so this is my 'good enough' solution to get started!