Create folder-contents.py
This script should be run from within a folder that contains project data from an Open Humans project. This expects there to be a list of folders named with the 8-digit project member ID. It looks in each numbered memberID folder; within the direct-sharing-* folder (there might be multiple based on the uploader types; and creates an output CSV file that lists the memberID and each individual file name.